It turns out the tests were only testing English values when 'English' was selected as a random language during th non english tests.
So I took the oppertunity to do some refactoring while I corrected the problem.

The tests now generate a english_config.ini file at the start of the test which
formats the string containing the flavor name to the flavor being installed.

All objects with label values that are being tested use a check_visible and check_hidden functions which removes alot of repetetive test code